feat: establish shared runtime workspace
