+++ .dockerignore
@@ -1,0 +1,26 @@
+.git/
+.github/
+.venv/
+**/.venv/
+__pycache__/
+**/__pycache__/
+*.py[cod]
+node_modules/
+**/node_modules/
+.next/
+**/.next/
+.env
+.env.*
+*.gguf
+services/summarize/models/
+.pytest_cache/
+**/.pytest_cache/
+.mypy_cache/
+**/.mypy_cache/
+.ruff_cache/
+**/.ruff_cache/
+coverage/
+htmlcov/
+terraform.tfstate
+*.tfstate
+*.tfstate.*
