--- .env.database.example
@@ -1,6 +1,0 @@
-# PostgreSQL Database Credentials
-# Used by: docker-compose.database.yml
-
-POSTGRES_USER=folium
-POSTGRES_PASSWORD=dev_password_change_in_prod
-POSTGRES_DB=folium_db
