+++ .env.example
@@ -1,28 +1,19 @@
-# Folium local runtime baseline. Copy this file to .env before running Compose.
-# A service may provide its own untracked .env to override only differing values.
+# Folium local secrets and optional overrides.
+# Compose owns non-sensitive local defaults. Copy to .env and set real values.
 
-APP_NAME=Folium
-LOG_LEVEL=INFO
-NEXT_PUBLIC_API_URL=http://localhost:8000
-
-# Host-run defaults. Compose overrides these with Docker DNS names where needed.
-BACKEND_API_URL=http://localhost:8000
-TRANSCRIPTION_SERVICE_URL=http://localhost:8001
-SUMMARIZATION_SERVICE_URL=http://localhost:8002
-TEMPORAL_ADDRESS=localhost:7233
-TEMPORAL_NAMESPACE=default
-VOICE_NOTE_TASK_QUEUE=voice-notes-queue
-
-# Replace placeholders in an untracked root or service-local .env.
-DATABASE_URL=postgresql+asyncpg://southdrift:change-me@southdrift-postgres:5432/southdrift_db
+# Application and database secrets
+DATABASE_URL=postgresql+asyncpg://folium:replace-with-a-local-password@folium-postgres:5432/folium_db
 JWT_SECRET=replace-with-a-local-secret
-STORAGE_PROVIDER=minio
-STORAGE_BUCKET=southdrift-dev
-STORAGE_REGION=us-east-1
+POSTGRES_PASSWORD=replace-with-a-local-postgres-password
+TEMPORAL_POSTGRES_PASSWORD=replace-with-a-local-temporal-postgres-password
+
+# Storage and object-store secrets
 STORAGE_ACCESS_KEY=replace-with-a-local-access-key
 STORAGE_SECRET_KEY=replace-with-a-local-secret-key
+MINIO_ROOT_USER=replace-with-a-local-minio-user
+MINIO_ROOT_PASSWORD=replace-with-a-local-minio-password
 
-# Provider credentials are intentionally unset.
+# Optional provider credentials; set only when the provider is enabled.
 # OPENAI_API_KEY=
 # AWS_ACCESS_KEY_ID=
 # AWS_SECRET_ACCESS_KEY=
