--- .env.storage.example
@@ -1,5 +1,0 @@
-# MinIO Storage Credentials
-# Used by: docker-compose.storage.yml
-
-MINIO_ROOT_USER=minioadmin
-MINIO_ROOT_PASSWORD=minioadmin
