+++ backend/.env.example
@@ -1,55 +1,5 @@
-# Application Settings
-APP_NAME=Folium
-DEBUG=True
-VERSION=1.0.0
+# Optional backend overrides. Root .env loads first.
+# Copy to backend/.env only when host-run backend values differ from the baseline.
 
-# CORS
-ALLOWED_ORIGINS=["http://localhost:3000", "https://localhost:3000"]
-
-# Database (future)
-DATABASE_URL=postgresql+asyncpg://folium:dev_password_change_in_prod@localhost:5432/folium_db
-
-# Authentication
-JWT_SECRET=CHANGE_THIS_LOCAL_DEV_SECRET
-
-# Multi-Cloud Storage Configuration
-# For local development with MinIO:
 STORAGE_ENDPOINT=http://localhost:9000
-STORAGE_ACCESS_KEY=minioadmin
-STORAGE_SECRET_KEY=minioadmin
-STORAGE_BUCKET=folium-dev
-STORAGE_REGION=us-east-1
-STORAGE_CDN_URL=
-
-# Transcription Service
-TRANSCRIPTION_SERVICE_URL=http://localhost:8001
-
-# For AWS S3 (production):
-# STORAGE_ENDPOINT=
-# STORAGE_ACCESS_KEY=YOUR_AWS_ACCESS_KEY
-# STORAGE_SECRET_KEY=YOUR_AWS_SECRET_KEY
-# STORAGE_BUCKET=folium-prod-us-east-1
-# STORAGE_REGION=us-east-1
-# STORAGE_CDN_URL=https://d111111abcdef8.cloudfront.net
-
-# For Azure Blob Storage (production):
-# STORAGE_ENDPOINT=https://youraccount.blob.core.windows.net
-# STORAGE_ACCESS_KEY=your_storage_account_name
-# STORAGE_SECRET_KEY=your_storage_account_key
-# STORAGE_BUCKET=documents
-# STORAGE_REGION=eastus
-# STORAGE_CDN_URL=https://youraccount.azureedge.net
-
-# Legacy Azure Storage (deprecated)
-AZURE_STORAGE_CONNECTION_STRING=
-AZURE_STORAGE_CONTAINER=uploads
-
-# AWS
-AWS_LAMBDA_ENDPOINT=
-AWS_REGION=us-east-1
-
-# Azure AI
-AZURE_FUNCTIONS_ENDPOINT=
-
-# Logging
-LOG_LEVEL=INFO
+STORAGE_PUBLIC_ENDPOINT=http://localhost:9000
