+++ backend/Dockerfile
@@ -1,53 +1,36 @@
 # Backend Dockerfile - FastAPI with uv for fast dependency management
-FROM python:3.11-slim as base
+FROM ghcr.io/astral-sh/uv:python3.13-bookworm-slim AS base
 
-# Install uv for fast package management
-COPY --from=ghcr.io/astral-sh/uv:latest /uv /usr/local/bin/uv
+WORKDIR /workspace
 
-# Set working directory
-WORKDIR /app
-
-# Set environment variables
 ENV PYTHONUNBUFFERED=1 \
     PYTHONDONTWRITEBYTECODE=1 \
-    UV_SYSTEM_PYTHON=1
+    PATH="/workspace/.venv/bin:$PATH"
 
 # Development stage
-FROM base as development
+FROM base AS development
 
-# Copy dependency files
-COPY pyproject.toml ./
-
-# Install dependencies with uv (much faster than pip)
-RUN uv pip install -e .[dev]
-
-# Copy application code
 COPY . .
+RUN uv sync --frozen --package folium --extra dev --group dev
 
 # Expose port
 EXPOSE 8000
 
-# Default command (can be overridden by docker-compose)
+WORKDIR /workspace/backend
 CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
 
 # Production stage
-FROM base as production
+FROM base AS production
 
-# Copy dependency files
-COPY pyproject.toml ./
-
-# Install dependencies
-RUN uv pip install .
-
-# Copy application code
 COPY . .
+RUN uv sync --frozen --no-dev --package folium
 
 # Create non-root user
-RUN useradd -m -u 1000 appuser && chown -R appuser:appuser /app
+RUN useradd -m -u 1000 appuser && chown -R appuser:appuser /workspace
 USER appuser
 
 # Expose port
 EXPOSE 8000
 
-# Production command (no reload)
+WORKDIR /workspace/backend
 CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
