+++ docker-compose.storage.yml
@@ -1,7 +1,7 @@
 services:
-  southdrift-minio:
-    image: minio/minio:latest
-    container_name: southdrift-minio
+  folium-minio:
+    image: minio/minio:RELEASE.2025-04-22T22-12-26Z
+    container_name: folium-minio
     ports:
       - "9000:9000" # S3 API
       - "9001:9001" # Web Console
