+++ folium.code-workspace
@@ -7,6 +7,9 @@
     {
       "path": "../pedropcamellon.github.io",
     },
+    {
+      "path": "../tasks",
+    },
   ],
   "settings": {
     // Global workspace settings
