+++ packages/folium-core/README.md
@@ -1,0 +1,4 @@
+# folium-core
+
+Shared Folium contracts and pure primitives. It intentionally excludes persistence,
+routes, provider integrations, and service-side effects.
