+++ packages/folium-core/pyproject.toml
@@ -1,0 +1,14 @@
+[project]
+name = "folium-core"
+version = "0.1.0"
+description = "Shared Folium contracts and pure primitives"
+readme = "README.md"
+requires-python = ">=3.13,<3.14"
+dependencies = []
+
+[build-system]
+requires = ["hatchling>=1.27.0"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/folium"]
