+++ packages/folium-core/src/folium/core/__init__.py
@@ -1,0 +1,1 @@
+"""Shared Folium contracts and pure primitives."""
