+++ packages/folium-core/src/folium/core/models/__init__.py
@@ -1,0 +1,5 @@
+"""Cross-project serializable data contracts."""
+
+from folium.core.models.workflow import AudioReference, VoiceNotesInput
+
+__all__ = ["AudioReference", "VoiceNotesInput"]
