+++ packages/folium-core/src/folium/core/models/workflow.py
@@ -1,0 +1,23 @@
+"""Serializable contracts shared by voice-note workflow callers and workers."""
+
+from dataclasses import dataclass
+
+
+@dataclass
+class AudioReference:
+    """Location and descriptive metadata for audio consumed by a workflow."""
+
+    storage_key: str
+    audio_url: str | None = None
+    bucket: str | None = None
+    original_filename: str | None = None
+    content_type: str | None = None
+
+
+@dataclass
+class VoiceNotesInput:
+    """Voice-note workflow input independent of persistence and transport layers."""
+
+    interaction_id: str
+    patient_id: str
+    audio: AudioReference
