+++ services/summarize/.env.example
@@ -1,0 +1,17 @@
+# Optional summarize-service overrides. Root .env loads first.
+# Copy to services/summarize/.env only when these values must differ locally.
+
+SUMMARIZATION_PROVIDER=local
+LOCAL_MODEL_PATH=/models/model.gguf
+LOCAL_MODEL_NAME=mediphi-clinical
+LOCAL_N_CTX=8192
+LOCAL_N_THREADS=4
+LOCAL_TEMPERATURE=0.0
+LOCAL_MAX_TOKENS=800
+
+# Required only when SUMMARIZATION_PROVIDER selects the corresponding provider.
+# OPENAI_API_KEY=
+# AWS_ACCESS_KEY_ID=
+# AWS_SECRET_ACCESS_KEY=
+# AZURE_OPENAI_ENDPOINT=
+# AZURE_OPENAI_KEY=
