+++ services/summarize/Dockerfile
@@ -1,15 +1,14 @@
-# Multi-stage Dockerfile for clinical summarization service
-# Supports multiple providers: local (default), openai, bedrock, azure
-
-# Build argument to select provider (default: local)
+# Clinical summarization service using the root uv workspace.
 ARG PROVIDER=local
 
-# ============================================================================
-# Base Stage: Common dependencies for all providers
-# ============================================================================
-FROM python:3.11-slim AS base
+FROM ghcr.io/astral-sh/uv:python3.13-bookworm-slim AS base
 
-WORKDIR /app
+ARG PROVIDER
+
+WORKDIR /workspace
+
+ENV PATH="/workspace/.venv/bin:$PATH" \
+    CMAKE_ARGS="-DGGML_NATIVE=OFF -DGGML_CPU_ALL_VARIANTS=OFF"
 
 # Install system dependencies
 RUN apt-get update && apt-get install -y \
@@ -17,58 +16,13 @@ RUN apt-get update && apt-get install -y \
     curl \
     && rm -rf /var/lib/apt/lists/*
 
-# Install uv for fast dependency management
-RUN pip install --no-cache-dir uv
+COPY . .
+RUN uv sync --frozen --no-dev --package summarize --extra "${PROVIDER}"
 
-# Copy dependency files and source code
-COPY pyproject.toml README.md ./
-COPY app ./app
-COPY prompts ./prompts
-
-# ============================================================================
-# Local LLM Stage: Install llama-cpp-python
-# ============================================================================
-FROM base AS provider-local
-
-# Install dependencies with local extras
-RUN uv pip install --system --no-cache ".[local]"
-
-# Copy MediPhi-Clinical model (from local cache or download manually)
-# If models/model.gguf exists locally, it will be copied. Otherwise:
-# Download manually: curl -L https://huggingface.co/mradermacher/MediPhi-Clinical-GGUF/resolve/main/MediPhi-Clinical.Q4_K_S.gguf -o models/model.gguf
+# Local models are mounted at runtime from ./services/summarize/models.
 RUN mkdir -p /models
-COPY models/model.gguf /models/model.gguf
 
-# ============================================================================
-# OpenAI Stage: Install OpenAI SDK
-# ============================================================================
-FROM base AS provider-openai
-
-# Install dependencies with openai extras from pyproject.toml
-RUN uv pip install --system --no-cache ".[openai]"
-
-# ============================================================================
-# AWS Bedrock Stage: Install boto3
-# ============================================================================
-FROM base AS provider-bedrock
-
-# Install dependencies with aws extras from pyproject.toml
-RUN uv pip install --system --no-cache ".[aws]"
-
-# ============================================================================
-# Azure OpenAI Stage: Install Azure SDK
-# ============================================================================
-FROM base AS provider-azure
-
-# Install dependencies with azure extras from pyproject.toml
-RUN uv pip install --system --no-cache ".[azure]"
-
-# ============================================================================
-# Final Stage: Select provider and copy application code
-# ============================================================================
-FROM provider-${PROVIDER} AS final
-
-WORKDIR /app
+WORKDIR /workspace/services/summarize
 
 # Expose service port
 EXPOSE 8002
@@ -78,5 +32,5 @@ HEALTHCHECK --interval=30s --timeout=10s --start-period=60s --retries=3 \
     CMD curl -f http://localhost:8002/health || exit 1
 
 # Run the application with auto-reload for development
-CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8002", "--reload"]
+CMD ["uv", "run", "--frozen", "--no-sync", "--package", "summarize", "uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8002", "--reload"]
