+++ services/summarize/pyproject.toml
@@ -3,7 +3,7 @@ name = "summarize"
 version = "0.1.0"
 description = "Clinical summarization microservice with Abstract Provider Pattern"
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.13,<3.14"
 dependencies = [
     "fastapi>=0.109.0",
     "uvicorn[standard]>=0.27.0",
