--- services/temporal/.env.example
@@ -1,32 +1,0 @@
-# =============================================================================
-# Temporal Infrastructure - Environment Variables
-# =============================================================================
-# Copy this file to .env and customize for your local environment
-#
-# Usage:
-#   cp .env.example .env
-#   # Edit .env with your values
-#   docker compose up
-# =============================================================================
-
-# PostgreSQL Configuration
-POSTGRES_DB=temporal
-POSTGRES_PASSWORD=temporal_dev_password_change_in_prod
-POSTGRES_SEEDS=southdrift-temporal-postgres
-POSTGRES_USER=temporal
-
-
-# Temporal Server Configuration
-# DB type (postgres12, postgres, mysql, cassandra)
-DB=postgres12
-DB_PORT=5432
-POSTGRES_PWD=temporal_dev_password_change_in_prod
-
-# Temporal UI Configuration
-TEMPORAL_UI_PORT=8080
-
-# Prometheus Metrics
-PROMETHEUS_ENDPOINT=0.0.0.0:9090
-
-# Optional: Custom namespace (default: default)
-# TEMPORAL_DEFAULT_NAMESPACE=default
