+++ services/transcribe/.env.example
@@ -1,22 +1,23 @@
-# ============== Provider Selection ==============
-# Choose ONE: whisper (self-hosted), aws, azure
+# Optional transcribe-service overrides. Root .env loads first.
+# Copy to services/transcribe/.env only when these values must differ locally.
+
+# Choose ONE: whisper (self-hosted), aws, azure.
 TRANSCRIPTION_PROVIDER=whisper
 
 # ============== Self-Hosted Whisper (HIPAA Compliant - No BAA Required) ==============
-WHISPER_MODEL_SIZE=base  # tiny|base|small|medium|large
-WHISPER_DEVICE=cpu       # cpu|cuda (GPU 4x faster)
+WHISPER_MODEL_SIZE=base
+WHISPER_DEVICE=cpu
 
-# ============== AWS Transcribe (Requires BAA for HIPAA) ==============
-# AWS_ACCESS_KEY_ID=your-access-key
-# AWS_SECRET_ACCESS_KEY=your-secret-key
+# Required only when TRANSCRIPTION_PROVIDER=aws.
+# AWS_ACCESS_KEY_ID=
+# AWS_SECRET_ACCESS_KEY=
 # AWS_REGION=us-east-1
-# TEMP_S3_BUCKET=southdrift-transcribe-temp
+# TEMP_S3_BUCKET=folium-transcribe-temp
 
-# ============== Azure Speech (Requires BAA for HIPAA) ==============
-# AZURE_SPEECH_KEY=your-subscription-key
+# Required only when TRANSCRIPTION_PROVIDER=azure.
+# AZURE_SPEECH_KEY=
 # AZURE_SPEECH_REGION=eastus
 
-# ============== Service Settings ==============
 PORT=8001
 LOG_LEVEL=info
 MAX_AUDIO_SIZE_MB=100
