+++ services/transcribe/.python-version
@@ -1,1 +1,1 @@
-3.11
+3.13
