+++ services/transcribe/pyproject.toml
@@ -3,7 +3,7 @@ name = "transcribe"
 version = "1.0.0"
 description = "Audio transcription microservice with Abstract Provider Pattern"
 readme = "README.md"
-requires-python = "==3.11.15"
+requires-python = ">=3.13,<3.14"
 dependencies = [
     "fastapi>=0.115.0",
     "uvicorn[standard]>=0.32.0",
