+++ services/voicenotes/.env.example
@@ -1,0 +1,6 @@
+# Optional voice-notes worker overrides. Root .env loads first.
+# Copy to services/voicenotes/.env only when these values must differ locally.
+
+# Required only when the backend enforces worker authentication.
+BACKEND_API_TOKEN=
+REQUEST_TIMEOUT_SECONDS=30
