+++ services/voicenotes/Dockerfile
@@ -1,13 +1,14 @@
-FROM python:3.11.15-slim
+FROM ghcr.io/astral-sh/uv:python3.13-bookworm-slim AS base
 
 ENV PYTHONDONTWRITEBYTECODE=1 \
-    PYTHONUNBUFFERED=1
+    PYTHONUNBUFFERED=1 \
+    PATH="/workspace/.venv/bin:$PATH"
 
-WORKDIR /app
+WORKDIR /workspace
 
-COPY pyproject.toml README.md ./
-COPY app ./app
+COPY . .
 
-RUN pip install --no-cache-dir .
+RUN uv sync --frozen --no-dev --package voicenotes
 
-CMD ["python", "-m", "app.worker"]
+WORKDIR /workspace/services/voicenotes
+CMD ["uv", "run", "--frozen", "--no-sync", "--package", "voicenotes", "python", "-m", "app.worker"]
