+++ services/voicenotes/docker-compose.service.yml
@@ -1,8 +1,10 @@
 services:
   folium-voicenotes-worker:
+    image: folium-voicenotes-worker
+    pull_policy: build
     build:
-      context: .
-      dockerfile: Dockerfile
+      context: ../..
+      dockerfile: services/voicenotes/Dockerfile
     container_name: folium-voicenotes-worker
     env_file:
       - path: ../../.env
@@ -10,32 +12,40 @@ services:
       - path: .env
         required: false
     depends_on:
-      southdrift-backend:
+      folium-backend:
         condition: service_healthy
       folium-summarize:
         condition: service_healthy
-      southdrift-temporal:
+      folium-temporal:
         condition: service_healthy
-      southdrift-transcribe:
+      folium-transcribe:
         condition: service_healthy
     environment:
-      - PYTHONUNBUFFERED=1
-      - TEMPORAL_ADDRESS=southdrift-temporal:7233
-      - TEMPORAL_NAMESPACE=default
-      - VOICE_NOTE_TASK_QUEUE=voice-notes-queue
-      - BACKEND_API_URL=http://southdrift-backend:8000
-      - TRANSCRIPTION_SERVICE_URL=http://southdrift-transcribe:8001
-      - SUMMARIZATION_SERVICE_URL=http://folium-summarize:8002
-      - LOG_LEVEL=INFO
+      PYTHONUNBUFFERED: "1"
+      TEMPORAL_ADDRESS: ${TEMPORAL_ADDRESS:-folium-temporal:7233}
+      TEMPORAL_NAMESPACE: ${TEMPORAL_NAMESPACE:-default}
+      VOICE_NOTE_TASK_QUEUE: ${VOICE_NOTE_TASK_QUEUE:-voice-notes-queue}
+      BACKEND_API_URL: ${BACKEND_API_URL:-http://folium-backend:8000}
+      TRANSCRIPTION_SERVICE_URL: ${TRANSCRIPTION_SERVICE_URL:-http://folium-transcribe:8001}
+      SUMMARIZATION_SERVICE_URL: ${SUMMARIZATION_SERVICE_URL:-http://folium-summarize:8002}
+      LOG_LEVEL: ${LOG_LEVEL:-INFO}
     restart: unless-stopped
     develop:
       watch:
         - action: sync
           path: ./app
-          target: /app/app
+          target: /workspace/services/voicenotes/app
           ignore:
             - __pycache__/
             - "*.pyc"
             - .pytest_cache/
         - action: rebuild
-          path: pyproject.toml
+          path: ./pyproject.toml
+        - action: sync
+          path: ../../packages/folium-core/src
+          target: /workspace/packages/folium-core/src
+          ignore:
+            - __pycache__/
+            - "*.pyc"
+        - action: rebuild
+          path: ../../packages/folium-core/pyproject.toml
